// generated/src/app/api-models/stage-auth-proc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthProc {
  batchId: string;
  transactionId: number;
  seqNo: number;
  fromDate: Date;
  thruDate: Date;
  codeListQual: string;
  authProcedure: string;
  authorizedQty: number;
  allowedAmt: number;
  status: string;
  statusDate: Date;
  statusReason: string;
  denialNotification: string;
  notificationDate: Date;
  priceSchedule: string;
  priceRegion: string;
  costProcedureCd: string;
  authorizedCost: number;
  requestedAmt: number;
  requestedQty: number;
  purchasePrice: number;
  medDefCode: string;
  toothNumber: string;
  ndcCode: string;
  qty: number;
  costProcedureCd1: string;
  toothNote: string;
  providerTrackingNo: string;
  providerEntityId: string;
  clrnghseTrackingNo: string;
  clrnghseEntityId: string;
  responseTrackingNo: string;
  responseEntityId: string;
  insertDatetime: Date;
}